timetable: Initialize zero timepoint
Before, the zero timepoint was initialized in the RobotFactory and in the TimetableManager. Now, it is initialized in the Timetable object. All timetables should have the same zero timepoint.
